moved pod tests to synchronous
