+++ debugutils/debugutils_suite_test.go
@@ -2,13 +2,16 @@ package debugutils
 
 import (
 	"context"
+	"os"
 	"testing"
 
+	"cloud.google.com/go/storage"
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/solo-io/go-utils/installutils/helmchart"
 	"github.com/solo-io/go-utils/installutils/kuberesource"
+	"google.golang.org/api/iterator"
 )
 
 func TestDebugutils(t *testing.T) {
@@ -41,6 +44,22 @@ var (
 		return nil
 	}, func(data []byte) {})
 
-	// _ = SynchronizedAfterSuite(func() {}, func() {
-	// })
+	_ = SynchronizedAfterSuite(func() {}, func() {
+		ctx := context.TODO()
+		client, err := storage.NewClient(ctx)
+		Expect(err).NotTo(HaveOccurred())
+		bucket := client.Bucket("go-utils-test")
+		obj := os.ExpandEnv("$BUILD_ID")
+		it := bucket.Objects(ctx, &storage.Query{
+			Prefix: obj,
+		})
+		for {
+			objAttrs, err := it.Next()
+			if err != nil && err == iterator.Done {
+				break
+			}
+			Expect(err).NotTo(HaveOccurred())
+			Expect(bucket.Object(objAttrs.Name).Delete(ctx)).NotTo(HaveOccurred())
+		}
+	})
 )
